cleaning up/renaming preview pieces
