+++ src/core/themeInstall.js
